Embed header directly in layout and use variables
This commit removes the header partial and adds the markup
directly to the base layout in order to use the site, description
and logo variables defined in _config.yml.

Those variables are not available in partials. We could pass them
explicitly in the {% %} include statement. But it seems less
convoluted to just scrap the partial.