Fix FAQ numbered list using markdown
The FAQ page is mostly a numbered list. For this purpose the better
readable markdown syntax should do just as well as HTML.

The page already seems to be formatted with markdown in mind,
although jekyll treats it as html.

This commit changes the extension of the FAQ page from .html to
.md so that jekyll knows how to build. For more infos see
https://jekyllrb.com/tutorials/orderofinterpretation/

This commit also updates the formatting syntax so that it renders
nicely. No changes to the contents are made.